pyproject.toml
[project]
name = "tradebot"
version = "0.1.0"
description = "Swing/long-term equities+crypto trading bot (paper-first) using Alpaca"
requires-python = ">=3.10"
dependencies = [
  "alpaca-py>=0.42.0",
  "pandas>=2.1",
  "numpy>=1.26",
  "pydantic>=2.6",
  "python-dotenv>=1.0",
  "PyYAML>=6.0",
  "rich>=13.7",
  "pytz>=2024.1",
  "lxml>=5.0",
  "fastapi>=0.110",
  "uvicorn>=0.27",
  "plotly>=5.22",
  "httpx>=0.27",
  "pyarrow>=16.0",
]



[project.scripts]
tradebot = "tradebot.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
